refactor(session): extract prompt tool resolution
